+++ Module 1/screen_mover.h
@@ -50,7 +50,7 @@ void adjustDifficulty();
 extern Line line1;
 extern Line line2;
 
-extern unsigned long int game_score;
-extern unsigned long int difficulty_counter;
+extern int game_score;
+extern int difficulty_counter;
 
 #endif /* SCREEN_MOVER_H_ */
